+++ purenessscopeserver/purenessscopeserver/PurenessScopeServer/Mail/base64.cpp
@@ -68,7 +68,7 @@ int base64_decode(unsigned char* out, const char* in, int out_size)
         }
     }
 
-    return dst - out;
+    return (int)(dst - out);
 }
 
 /*****************************************************************************
